requirements.txt
absl-py==2.3.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.9
aiohttp-cors==0.8.1
aiosignal==1.3.1
ale-py==0.10.1
alembic==1.16.1
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
attrs==24.2.0
autopage==0.5.2
boto3==1.38.36
botocore==1.38.36
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.0
choreographer==1.0.9
click==8.2.1
cliff==4.10.0
cloudpickle==2.2.1
cmaes==0.11.1
cmd2==2.6.1
colorama==0.4.6
colorful==0.5.6
colorlog==6.9.0
contourpy==1.3.2
cycler==0.12.1
distlib==0.3.9
etils==1.12.2
Farama-Notifications==0.0.4
filelock==3.16.1
fonttools==4.58.1
frozenlist==1.5.0
fsspec==2025.5.1
glfw==2.9.0
google-api-core==2.25.1
google-auth==2.40.3
googleapis-common-protos==1.70.0
greenlet==3.2.3
grpcio==1.72.1
gymnasium==1.0.0
hydra-core==1.3.2
hydra-optuna-sweeper==1.2.0
hydra-ray-launcher==1.2.1
idna==3.10
imageio==2.37.0
importlib_metadata==8.7.0
importlib_resources==6.5.2
jax-jumpy==1.0.0
Jinja2==3.1.4
jmespath==1.0.1
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
kaleido==1.0.0
kiwisolver==1.4.8
logistro==1.1.0
Mako==1.3.10
Markdown==3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.9.2
mdurl==0.1.2
mpmath==1.3.0
msgpack==1.1.1
mujoco==3.3.2
multidict==6.1.0
munkres==1.1.4
narwhals==2.0.1
networkx==3.4.2
numpy==1.26.4
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
omegaconf==2.3.0
opencensus==0.11.4
opencensus-context==0.1.3
opencv-python==4.11.0.86
opentelemetry-api==1.34.1
opentelemetry-exporter-prometheus==0.55b1
opentelemetry-proto==1.34.1
opentelemetry-sdk==1.34.1
opentelemetry-semantic-conventions==0.55b1
optuna==2.10.1
orjson==3.11.1
packaging==25.0
pandas==2.2.3
pbr==6.1.1
pillow==11.2.1
platformdirs==4.3.8
plotly==6.2.0
ply==3.11
prettytable==3.16.0
prometheus_client==0.22.1
propcache==0.2.1
proto-plus==1.26.1
protobuf==5.29.5
psutil==6.1.0
py-spy==0.4.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.7
pydantic_core==2.33.2
pygame==2.6.1
Pygments==2.18.0
PyOpenGL==3.1.1a1
pyparsing==3.2.3
pyperclip==1.9.0
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
ray==2.47.0
referencing==0.36.2
requests==2.32.3
rich==13.9.4
rpds-py==0.25.1
rsa==4.9.1
s3transfer==0.13.0
scipy==1.15.3
simplejson==3.20.1
sip==6.12.0
six==1.17.0
smart-open==7.1.0
SQLAlchemy==2.0.41
stable_baselines3==2.6.0
stevedore==5.4.1
sympy==1.13.1
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tomli==2.2.1
torch==2.5.1
torch-geometric==2.6.1
tornado==6.5.1
tqdm==4.67.1
triton==3.1.0
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2024.2
unicodedata2==16.0.0
urllib3==2.2.3
virtualenv==20.31.2
wcwidth==0.2.13
Werkzeug==3.1.3
wrapt==1.17.2
yarl==1.18.3
zipp==3.22.0
